refactor(ProductFixture.test) KitchenposFixture.java -> ProductFixture 이름변경
